Serialize incoming events
All incoming events are funneled into a channel that is split to
per-account channels on read to prevent locking up while keeping
order of events consistent for each account.

When opening an account the user is required to call Open() to
open the events channel for it or it's possible some events are
dropped before StartIO during account configuration.

Previously event order was unknown because they were waiting in
separate parallel goroutines.